find_free_dirent

+++ fat32_fs/src/fat.rs
@@ -122,7 +122,7 @@ impl FAT {
     }
 
     // 获取start_cluster所在簇链中，从start_cluster开始的第index个簇
-    pub fn get_cluster_at(&self, start_cluster: u32, index: u32, block_device: &Arc<dyn BlockDevice>) -> u32 {
+    pub fn get_cluster_at(&self, start_cluster: u32, index: usize, block_device: &Arc<dyn BlockDevice>) -> u32 {
         let mut cluster = start_cluster;
         for _ in 0..index {
             assert!(cluster < self.max_cluster, "The current cluster number {} exceeds the maximum cluster number {}!"
